ref: extract plugin archive extraction and check runner to pkg
